Split up Constraints reducer
