+++ plugins/services/src/js/reducers/JSONAppReducers.js
@@ -1,4 +1,4 @@
-import {JSONReducer as constraints} from './serviceForm/Constraints';
+import {JSONReducer as constraints} from './serviceForm/JSONReducers/Constraints';
 import {JSONReducer as container} from './serviceForm/Container';
 import {JSONReducer as env} from './serviceForm/EnvironmentVariables';
 import {JSONReducer as fetch} from './serviceForm/JSONReducers/Artifacts';
