+++ plugins/services/src/js/reducers/JSONSingleContainerParser.js
@@ -1,5 +1,5 @@
 import {JSONParser as container} from './serviceForm/Container';
-import {JSONParser as constraints} from './serviceForm/Constraints';
+import {JSONParser as constraints} from './serviceForm/JSONReducers/Constraints';
 import {JSONParser as fetch} from './serviceForm/JSONReducers/Artifacts';
 import {JSONParser as environmentVariables} from './serviceForm/EnvironmentVariables';
 import {JSONParser as externalVolumes} from './serviceForm/ExternalVolumes';
