+++ plugins/services/src/js/reducers/serviceForm/FormReducers/Constraints.js
@@ -1,0 +1,7 @@
+import {
+  FormReducer
+} from '../common/Constraints';
+
+module.exports = {
+  FormReducer
+};
